requirements.txt
requests==2.32.3
lxml==5.3.0
Jinja2==3.1.4
PyYAML==6.0.2
beautifulsoup4==4.12.3
tenacity==9.0.0
python-dateutil==2.9.0.post0
## minihtml removed (unused)
pdfminer.six==20240706
markdown==3.7
PySocks==1.7.1
ocrmypdf==16.8.0
Pillow==10.4.0

# Dev / Test
pytest==8.3.3
pytest-cov==5.0.0
ruff==0.6.9
black==24.8.0
isort==5.13.2
python-dotenv==1.0.1
flask==3.0.3
